Change to use a hash of the keyfile instead of the contents of the keyfile itself
Before, everything after the first 64 bytes of the keyfile was discarded when generating an encryption key with it. If the keyfile was used alone, it was used raw, and if not, it was ran through HKDF along with the key derived from the password to derive a key.

Now, instead, the full key file is hashed and that hash is ran through HKDF as well. If the keyfile is used without a password, this is done to ensure that the proper key size is used in case the user selects a message digest that would produce a hash smaller than EVP_MAX_KEY_LENGTH.