Add backward compatibility for algorithms depcrecated in OpenSSl 3.0
OpenSSL 3.0 depcreates a number of older algorithms, such as whirlpool and bowfish. However instead of simply marking them depcreated, they do not work at all unless loaded through OpenSSL's new "provider" system. This uses a preprocessor macro test to see if the version of OpenSSL than or equal to 3 in order to load the legacy algorithms.