Rearranged source files
Before the way the source files were setup was a bit unorthodox and required a special flag to compile because of multiple definitions. This stuffs the things that need to be forward-declared into lib.h and defines them in lib.c, which is the more conventional way to do it.

Also modified the autotools package structure to put all of the object files into ./src/ and compile the binaries into ./bin/ for a cleaner top directory. As well, put the testing scripts into a scripts directory, and also made sure that it cleans up the testfiles and keyfiles once done.