Added cryptoheader to encrpted files and adjusted tooltips
Previously, a user would need to remember the parameters chosen for their encryption/digest algorithms as well as scrypt work factors to decrypt properly. I added a cryptoheader storing this information instead so that decryption can be done without remembering them. This also tests that the file being decrypted is actually a file encrypted with evpencutil before proceeding.

Also added some guidance to the Encryption Algorithm and Message Digest Algorithm tool-tip text boxes in the GUI.