Changed all malloc calls to calloc
Mostly done for consistency, since I generally think that memory initialized to 0 is the safest to work with. There are certain implementation-defined things that happen between malloc and calloc that might make calloc preferable as well, such as with Linux's out-of-memory management and malloc basically returning non-null 100% of the time whether there is sufficient memory for allocation or not. Calloc is just a little safer it seems.