Added global struct to use with atexit and changed test scripts to direct ASAN not to report memory leaks
I decided that I wanted to register the sanitiztion of sensitive buffers in cryptSt with a function registered to atexit, so I needed to include a global structure to point to those areas in memory. This also meant leaving the buffers unfreed at exit, which ASAN reports as leaked memory, but the OS frees.