Added ability to be prompted for password using the CLI instead of having to supply it on the command-line
Taking care of issue https://github.com/kennbr34/evpencutil/issues/1

I decided instead of prompting for a password in the absence of '-p' that I would just prompt for a password if user enters '-p prompt'. This is because the code to use a keyfile along with a password would get overcomplicated if the funcitonality was done by ommitting '-p', and since this should be primarily a GUI based application and the CLI is more of a supplementary, I think using '-p prompt' to be prompted is more than apt.