Added options to be prompted for a password instead of supplying it via command-line argument, as well as giving options to display it as typed, and/or verify it
Solved https://github.com/kennbr34/evpencutil/issues/2

Now a user can specify '-P' to be prompted for a password via the CLI intead of supplying it as a command line argument. User can also use '-V' to enable a verification step, or '-D' to display the password as it is typed at the prompt, to prevent typos.