Enabled pasisng keyfile by standard input, and fixed hashing of it
Firstly, the keyfile was being hashed in its entirety, instead of only the portion of it which is being used by the cipher. OpenSSL defines EVP_MAX_KEY_LENGTH and so limiting the hashing to that size is best.

Secondly, in case a uer needs/wants to pass the keyfile by standard input, that is now possible.