Fixed key-file-hashing mistake
I changed the key-file hash to only count the first EVP_MAX_KEY_LENGTH bytes of the file, but remembered that I specifically intended to rely on a hash of the entire file as part of the key generation. That way someone will STILL need the entire key file, not just the first part of it.