Stdio processing, and better sequence-enforcement and truncation-prevention (#5)
* Upload of 0.2.0

Added support for processing standard input/output

* Fixed formatting

* Stdio processing and chunked encryption improvements

This achieves the same stdio processing as the previous commit, but better handles the chunked encryption. The previous commit would not detect or prevent truncation of a message, and only used a sequentially incremented ID for each chunk to enforce sequence. Instead I adopted the strategy used by Monocphyer, which they call a "ratchet" where the authentication key for each chunk depends on the previous chunk of data. I achieve the same thing by making each call to HMAC on a chunk of cipher-text use a key that is derived from the previous chunk's MAC; that way the new HMAC key is dependent upon the cipher-text and its correct sequence. However, an 8-bit flag needed to be set to '1' to denote when the last chunk is being processed to avoid truncation.

Also fixed a bug I had used a quick workaround for, but was actually causing the very last chunk not to be authenticated. This was because I was only checking the MAC if 'remainingBytes' was not zero, which ultimately lead back to how I was reading the last chunk in with fread during decryption.