Add chunk sequencing and truncation protection, as well as different key geeration per chunk
This version generatins a new HMAC key that is based on the previous one for every next chunk's MAC computation. That key is also used to derive a different EVP key and IV pair for each chunk. That will enforce the sequencing of the chunks, and then adding the buffer size into the MAC computation will also ensure that truncation will be detected.