Parllelization testing
Further testing with parallelization show some promise, but also some difficult-to-grasp bugs. CBC modes once again seem to no longer work, and there also appears to be something no longer working with all associated data being able to be authenticated.

Testing with asan and valgrind, it seemed like there are some issues with memory addresses in the pthreads functions themselves, but that must be due to something about the inputs I am giving it.