Fixed padding issue for block ciphers
The conditional test to handle padding for the last block relied on the remainingBytes variable, however this updated to the same value for every thread. Instead, the original buffer size is tracked so that when st->cryptSt.fileBufSize is less than the original buffer size, it is a positive indicator that the last chunk of encryption is being processed.