Optimized mutex locking
Previous, the mutex lock was guarding the entirey of the code in the thread_*crypt_cunk() functions and wasn't allowing very good performance. I moved locks to only guard parts of the code that were making an assignment to a variable that controlls data flow, shared variables and write operations. The EVP library functions do not need locking because they are already thread safe.