Added features for easier benchmarking
The bencmarking scripts rely on a list of ciphers, and there should be an easy function to produce that list. I have implemnted a "list-all" and "list-supported" option when specifying a cipher with -c or a digest with -m respectively, so that "-c list-all" will produce a list of all ciphers, and "-m list-supported", would list only supported digests, etc.