First refactored attempt
This version is a little bit different than the previous multi-threaded attempts. The variables which are sent to the thread function are more isolated, and I have tried to separate their nitiation with thread creation to help reduce race conditions. While this seemed to work some, message authentication is still for some reason broken, and there are still intermittent bad-decryption issues.