Last multi-threading attempt
So far I have still not been able to eliminate all race conditions from the multi-threaded version of the program. I believe the approach I am taking with processing an unknown file-size and enrypting in chunks, along with the specialized key generation routine, may not be compatible with this multi-threading approach and I'd need to take other measures. This is especially true because after eliminating race conditions and synchronization issues with mutexes and conditional variables, the paralell version is no faster than the serialized version, and not all race conditions were able to be eliminated anyway.